Add unit tests for new launchers + small refactoring
